feat: animation selection and warning fix

+++ src/LThigh.cpp
@@ -33,7 +33,7 @@ void LThigh::draw(Animation &anim, Shader &ourShader, float newx, float newy)
     {
         new LLowerPart(_bodyStack, _model)
     };
-    for (int i = 0; i < parts.size(); i++)
+    for (size_t i = 0; i < parts.size(); i++)
     {
         _bodyStack->push(parts[i]);
         _bodyStack->top()->draw(anim, ourShader, x - 0.4f, y - .3125f);
